+++ main.py
@@ -3,68 +3,17 @@ import os
 import numpy as np
 import pygame
 
-from pycarphysics import VehicleEntity
-from pycarphysics.collisions.low import collide
-from pycarphysics.collisions.shapes import SquareShape, RectangleShape, PolygonShape
-from pycarphysics.piece.chassis import Chassis
-from pycarphysics.piece.steering import Steering
+from pycarphysics.collisions import filter_all_collisions
+from pycarphysics.collisions.response import push
+from pycarphysics.collisions.shapes import SquareShape, RectangleShape
+from pycarphysics.entities import VehicleEntity
+from pycarphysics.entities.chassis import Chassis
+from pycarphysics.entities.steering import Steering
 
 SCREEN_TITLE = "Car demo physic"
 SCREEN_SIZE = 640, 480
 DEFAULT_FRAME_RATE = 120
 CAR_IMAGE_FILENAME = "assets/car_image.png"
-
-
-def filter_all_collisions(entity: VehicleEntity, others: list[PolygonShape] | tuple[PolygonShape, ...]):
-    for other in others:
-        is_collide, move = collide(entity.points, other.points)
-
-        if is_collide:
-            yield other, move
-
-
-def slide(entity: VehicleEntity, other: PolygonShape, move: np.ndarray, **kwargs):
-    force = kwargs.get("force", 0.35)
-    deceleration = kwargs.get("deceleration", 0.25)
-
-    entity.collider.translate(move)
-
-    entity.chassis.position += move / entity.ppu
-    entity.motor.acceleration *= -deceleration
-    entity.velocity[0] *= -1 * force
-    entity.velocity[1] *= -1 * force
-
-    return entity, other
-
-
-def bounce(entity: VehicleEntity, other: PolygonShape, move: np.ndarray, **kwargs):
-    force = kwargs.get("force", 0.35)
-
-    entity.collider.translate(move)
-
-    entity.chassis.position += move / entity.ppu
-
-    entity.velocity[0] *= -1 * force
-    entity.velocity[1] *= -1 * force
-
-    return entity, other
-
-
-def push(entity: VehicleEntity, other: PolygonShape, move: np.ndarray, **kwargs):
-    force = kwargs.get("force", 0.95)
-    deceleration = kwargs.get("deceleration", 0.05)
-
-    entity.collider.translate(move)
-
-    entity.chassis.position += move / entity.ppu
-    entity.motor.acceleration *= deceleration
-
-    other.translate(-move)
-
-    entity.velocity[0] *= force
-    entity.velocity[1] *= force
-
-    return entity, other
 
 
 if __name__ == '__main__':
@@ -115,7 +64,7 @@ if __name__ == '__main__':
         box1.rotate(0.025 * dt)
 
         for box, move in filter_all_collisions(entity, (box0, box1)):
-            entity, box = push(entity, box, move)
+            entity, _ = push(entity, box, move)
 
         screen.fill((255, 255, 255))
         car_image_rotate = pygame.transform.rotate(car_image, angle)
